refactor(OutputFormmater): 출력 포멧 클래스 생성
